Refact visualize diagrams and preferences manager
Related to #536 and #560.
