+++ dev/plugins/hu.elte.txtuml.diagnostics/META-INF/MANIFEST.MF
@@ -15,6 +15,7 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.papyrus.infra.gmfdiag.css,
  hu.elte.txtuml.export.uml2,
  hu.elte.txtuml.utils,
+ hu.elte.txtuml.utils.eclipse,
  hu.elte.txtuml.export.papyrus,
  hu.elte.txtuml.api.model.execution
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
