+++ dev/plugins/hu.elte.txtuml.export.javascript/plugin.xml
@@ -6,14 +6,14 @@
    <extension
          point="org.eclipse.ui.commands">
       <command
-            defaultHandler="hu.elte.txtuml.export.javascript.handlers.TxtUMLVisualizeHandler"
-            id="hu.elte.txtuml.export.javascript.txtUMLVisualize"
-            name="txtUMLVisualize">
+            defaultHandler="hu.elte.txtuml.export.javascript.handlers.JavaScriptVisualizeHandler"
+            id="hu.elte.txtuml.export.javascript.JavaScriptVisualize"
+            name="JavaScriptVisualize">
       </command>
   	  <command
-            defaultHandler="hu.elte.txtuml.export.javascript.handlers.TxtUMLVisualizeSelectedDiagramsHandler"
-            id="hu.elte.txtuml.export.javascript.txtUMLVisualizeSelectedDiagram"
-            name="txtUMLVisualizeSelectedDiagram">
+            defaultHandler="hu.elte.txtuml.export.javascript.handlers.JavaScriptVisualizeSelectedDiagramsHandler"
+            id="hu.elte.txtuml.export.javascript.JavaScriptVisualizeSelectedDiagram"
+            name="JavaScriptVisualizeSelectedDiagram">
       </command>
    </extension>
    <extension
@@ -25,7 +25,7 @@
                id="hu.elte.txtuml.export.menu"
                label="txtUML">
             <command
-                  commandId="hu.elte.txtuml.export.javascript.txtUMLVisualize"
+                  commandId="hu.elte.txtuml.export.javascript.JavaScriptVisualize"
                   icon="icons/jsdiagram.png"
                   label="Generate JointJS diagrams from txtUML"
                   style="push">
@@ -36,7 +36,7 @@
 	  <menuContribution 
 	     locationURI="popup:org.eclipse.ui.navigator.ProjectExplorer#PopupMenu?before=additions">
 	     <command
-            commandId="hu.elte.txtuml.export.javascript.txtUMLVisualizeSelectedDiagram"
+            commandId="hu.elte.txtuml.export.javascript.JavaScriptVisualizeSelectedDiagram"
             icon="icons/jsdiagram.png"
             label="Visualize as JointJS diagram"
             style="push">
@@ -54,7 +54,7 @@
    <menuContribution
          locationURI="popup:org.eclipse.jdt.ui.PackageExplorer?before=additions">
       <command
-            commandId="hu.elte.txtuml.export.javascript.txtUMLVisualizeSelectedDiagram"
+            commandId="hu.elte.txtuml.export.javascript.JavaScriptVisualizeSelectedDiagram"
             icon="icons/jsdiagram.png"
             label="Visualize as JointJS diagram"
             style="push">
