+++ dev/plugins/hu.elte.txtuml.export.javascript/src/hu/elte/txtuml/export/javascript/handlers/JavaScriptVisualizeHandler.java
@@ -1,16 +1,16 @@
-package hu.elte.txtuml.export.papyrus.handlers;
+package hu.elte.txtuml.export.javascript.handlers;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
 import org.eclipse.jface.wizard.WizardDialog;
 
-import hu.elte.txtuml.export.papyrus.wizardz.TxtUMLVisualizeWizard;
+import hu.elte.txtuml.export.javascript.wizardz.JavaScriptVisualizeWizard;
 
 /**
  * Handles the call to txtUML Visualization
  */
-public class TxtUMLVisualizeHandler  extends AbstractHandler {
+public class JavaScriptVisualizeHandler extends AbstractHandler {
 
 	/**
 	 * Opens a TxtUMLVisuzalize wizard
@@ -18,12 +18,10 @@ public class TxtUMLVisualizeHandler  extends AbstractHandler {
 	@Override
 	public Object execute(ExecutionEvent event) throws ExecutionException {
 
+		WizardDialog wizardDialog = new WizardDialog(null, new JavaScriptVisualizeWizard());
 
-		WizardDialog wizardDialog = new WizardDialog( null,
-	      new TxtUMLVisualizeWizard());
-		
 		wizardDialog.open();
 		return null;
 	}
-	
+
 }
