+++ dev/plugins/hu.elte.txtuml.export.javascript/src/hu/elte/txtuml/export/javascript/wizardz/JavaScriptVisualizeWizard.java
@@ -2,31 +2,25 @@ package hu.elte.txtuml.export.javascript.wizardz;
 
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.jdt.core.IType;
-import org.eclipse.jdt.core.JavaModelException;
 import org.eclipse.jface.operation.IRunnableWithProgress;
-import org.eclipse.jface.wizard.Wizard;
 import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.progress.IProgressService;
 
+import hu.elte.txtuml.api.layout.ClassDiagram;
+import hu.elte.txtuml.api.layout.CompositeDiagram;
+import hu.elte.txtuml.api.layout.StateMachineDiagram;
 import hu.elte.txtuml.export.javascript.Exporter;
 import hu.elte.txtuml.export.papyrus.layout.txtuml.TxtUMLExporter;
 import hu.elte.txtuml.export.papyrus.layout.txtuml.TxtUMLLayoutDescriptor;
-import hu.elte.txtuml.export.papyrus.preferences.PreferencesManager;
-import hu.elte.txtuml.export.papyrus.wizardz.VisualizeTxtUMLPage;
 import hu.elte.txtuml.export.uml2.ExportMode;
 import hu.elte.txtuml.export.uml2.TxtUMLToUML2;
 import hu.elte.txtuml.layout.export.DiagramExportationReport;
@@ -34,89 +28,33 @@ import hu.elte.txtuml.utils.Logger;
 import hu.elte.txtuml.utils.Pair;
 import hu.elte.txtuml.utils.eclipse.Dialogs;
 import hu.elte.txtuml.utils.eclipse.SaveUtils;
-import hu.elte.txtuml.utils.eclipse.WizardUtils;
+import hu.elte.txtuml.utils.eclipse.preferences.PreferencesManager;
+import hu.elte.txtuml.utils.eclipse.wizards.TxtUMLVisualizeWizard;
+import hu.elte.txtuml.utils.eclipse.wizards.VisualizeTxtUMLPage;
 
 /**
  * Wizard for visualization of txtUML models
  */
-public class TxtUMLVisualizeWizard extends Wizard {
+public class JavaScriptVisualizeWizard extends TxtUMLVisualizeWizard {
 
-	private VisualizeTxtUMLPage selectTxtUmlPage;
-
-	/**
-	 * The Constructor
-	 */
-	public TxtUMLVisualizeWizard() {
+	public JavaScriptVisualizeWizard() {
 		super();
 		setNeedsProgressMonitor(true);
 	}
 
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see org.eclipse.jface.wizard.Wizard#getWindowTitle()
-	 */
 	@Override
 	public String getWindowTitle() {
 		return "Create Javascript Diagrams from txtUML Model";
 	}
 
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see org.eclipse.jface.wizard.Wizard#addPages()
-	 */
 	@Override
 	public void addPages() {
-		selectTxtUmlPage = new VisualizeTxtUMLPage();
+		selectTxtUmlPage = new VisualizeTxtUMLPage(StateMachineDiagram.class, ClassDiagram.class, CompositeDiagram.class);
 		addPage(selectTxtUmlPage);
 	}
 
-	/**
-	 * Calls the {@link hu.elte.txtuml.export.uml2.UML2 txtUML UML2 Export} and
-	 * then starts the visualization.
-	 */
 	@Override
-	public boolean performFinish() {
-		List<IType> txtUMLLayout = selectTxtUmlPage.getTxtUmlLayouts();
-		Map<Pair<String, String>, List<IType>> layoutConfigs = new HashMap<>();
-		List<String> invalidLayouts = new ArrayList<>();
-		for (IType layout : txtUMLLayout) {
-			Optional<Pair<String, String>> maybeModel = Optional.empty();
-			try {
-				maybeModel = Stream.of(layout.getTypes())
-						.map(innerClass -> WizardUtils.getModelByAnnotations(innerClass)).filter(Optional::isPresent)
-						.map(Optional::get).findFirst();
-			} catch (JavaModelException e) {
-				Logger.sys.error(e.getMessage());
-				return false;
-			}
-
-			if (maybeModel.isPresent()) {
-				Pair<String, String> model = maybeModel.get();
-				if (!layoutConfigs.containsKey(model)) {
-					layoutConfigs.put(model, new ArrayList<>(Arrays.asList(layout)));
-				} else {
-					layoutConfigs.get(model).add(layout);
-				}
-			} else {
-				invalidLayouts.add(layout.getElementName());
-			}
-		}
-
-		if (!invalidLayouts.isEmpty()) {
-			Dialogs.MessageBox("Invalid layouts", "The following diagram descriptions have no txtUML model attached"
-					+ ", hence no diagram is generated for them:" + System.lineSeparator() + invalidLayouts.stream()
-							.map(s -> " - ".concat(s)).collect(Collectors.joining(System.lineSeparator())));
-		}
-
-		PreferencesManager.setValue(PreferencesManager.TXTUML_VISUALIZE_TXTUML_LAYOUT, layoutConfigs.values().stream()
-				.flatMap(c -> c.stream()).map(layout -> layout.getFullyQualifiedName()).collect(Collectors.toList()));
-
-		PreferencesManager.setValue(PreferencesManager.TXTUML_VISUALIZE_TXTUML_LAYOUT_PROJECTS,
-				layoutConfigs.values().stream().flatMap(c -> c.stream())
-						.map(layout -> layout.getJavaProject().getElementName()).collect(Collectors.toList()));
-
+	protected boolean exportDiagrams(Map<Pair<String, String>, List<IType>> layoutConfigs, List<IType> txtUMLLayout) {
 		for (Pair<String, String> model : layoutConfigs.keySet()) {
 			String txtUMLModelName = model.getFirst();
 			String txtUMLProjectName = model.getSecond();
@@ -219,19 +157,6 @@ public class TxtUMLVisualizeWizard extends Wizard {
 			}
 		}
 		return true;
-	}
-
-	private void checkNoLayoutDescriptionsSelected() throws InterruptedException {
-		if (selectTxtUmlPage.getTxtUmlLayouts().isEmpty()) {
-			boolean answer = Dialogs.WarningConfirm("No Layout descriptions",
-					"No diagrams will be generated using the current setup,"
-							+ " because no diagram descriptions are added." + System.lineSeparator()
-							+ "In order to have diagrams visualized, select a description from the wizard."
-							+ System.lineSeparator() + System.lineSeparator()
-							+ "Do you want to continue without diagram descriptions?");
-			if (!answer)
-				throw new InterruptedException();
-		}
 	}
 
 }
