+++ dev/plugins/hu.elte.txtuml.export.plantuml/META-INF/MANIFEST.MF
@@ -16,8 +16,7 @@ Require-Bundle: hu.elte.txtuml.api.model,
  org.eclipse.ui.workbench,
  org.eclipse.emf.ecore,
  org.eclipse.ui,
- org.eclipse.core.expressions,
- hu.elte.txtuml.export.papyrus
+ org.eclipse.core.expressions
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Vendor: ELTE
 Export-Package: hu.elte.txtuml.export.plantuml,
