+++ dev/plugins/hu.elte.txtuml.export.plantuml/src/hu/elte/txtuml/export/plantuml/wizards/PlantUMLVisualizeWizard.java
@@ -1,114 +1,49 @@
 package hu.elte.txtuml.export.plantuml.wizards;
 
 import java.lang.reflect.InvocationTargetException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.jdt.core.IType;
-import org.eclipse.jdt.core.JavaModelException;
 import org.eclipse.jface.operation.IRunnableWithProgress;
-import org.eclipse.jface.wizard.Wizard;
 import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.progress.IProgressService;
 
-import hu.elte.txtuml.export.papyrus.preferences.PreferencesManager;
+import hu.elte.txtuml.api.model.seqdiag.SequenceDiagram;
 import hu.elte.txtuml.export.plantuml.PlantUmlExporter;
 import hu.elte.txtuml.export.plantuml.exceptions.SequenceDiagramExportException;
 import hu.elte.txtuml.utils.Logger;
 import hu.elte.txtuml.utils.Pair;
 import hu.elte.txtuml.utils.eclipse.Dialogs;
 import hu.elte.txtuml.utils.eclipse.SaveUtils;
-import hu.elte.txtuml.utils.eclipse.WizardUtils;
+import hu.elte.txtuml.utils.eclipse.preferences.PreferencesManager;
+import hu.elte.txtuml.utils.eclipse.wizards.TxtUMLVisualizeWizard;
+import hu.elte.txtuml.utils.eclipse.wizards.VisualizeTxtUMLPage;
 
-public class PlantUMLVisualizeWizard extends Wizard {
+public class PlantUMLVisualizeWizard extends TxtUMLVisualizeWizard {
 
-	private VisualizePlantUMLPage selectPlantUmlPage;
-
-	/**
-	 * The Constructor
-	 */
 	public PlantUMLVisualizeWizard() {
 		super();
 		setNeedsProgressMonitor(true);
 	}
-
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see org.eclipse.jface.wizard.Wizard#getWindowTitle()
-	 */
+	
 	@Override
 	public String getWindowTitle() {
-		return "Create Papyrus model from txtUML model";
+		return "Create PlantUML diagram from txtUML sequence diagram";
 	}
-
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see org.eclipse.jface.wizard.Wizard#addPages()
-	 */
+	
 	@Override
 	public void addPages() {
-		selectPlantUmlPage = new VisualizePlantUMLPage();
-		addPage(selectPlantUmlPage);
+		selectTxtUmlPage = new VisualizeTxtUMLPage(SequenceDiagram.class);
+		addPage(selectTxtUmlPage);
 	}
 
-	/**
-	 * Calls the {@link hu.elte.txtuml.export.uml2.UML2 txtUML UML2 Export} and
-	 * then starts the visualization.
-	 */
 	@Override
-	public boolean performFinish() {
-		List<IType> txtUMLLayout = selectPlantUmlPage.getPlantUmlLayouts();
-		Map<Pair<String, String>, List<IType>> layoutConfigs = new HashMap<>();
-		List<String> invalidLayouts = new ArrayList<>();
-		for (IType layout : txtUMLLayout) {
-			Optional<Pair<String, String>> maybeModel = Optional.empty();
-			try {
-				maybeModel = Stream.of(layout.getTypes())
-						.map(innerClass -> WizardUtils.getModelByAnnotations(innerClass)).filter(Optional::isPresent)
-						.map(Optional::get).findFirst();
-				if (!maybeModel.isPresent())
-					maybeModel = WizardUtils.getModelByFields(layout);
-			} catch (JavaModelException e) {
-				Logger.sys.error(e.getMessage());
-				return false;
-			}
-
-			if (maybeModel.isPresent()) {
-				Pair<String, String> model = maybeModel.get();
-				if (!layoutConfigs.containsKey(model)) {
-					layoutConfigs.put(model, new ArrayList<>(Arrays.asList(layout)));
-				} else {
-					layoutConfigs.get(model).add(layout);
-				}
-			} else {
-				invalidLayouts.add(layout.getElementName());
-			}
-		}
-
-		if (!invalidLayouts.isEmpty()) {
-			Dialogs.MessageBox("Invalid layouts", "The following diagram descriptions have no txtUML model attached"
-					+ ", hence no diagram is generated for them:" + System.lineSeparator() + invalidLayouts.stream()
-							.map(s -> " - ".concat(s)).collect(Collectors.joining(System.lineSeparator())));
-		}
-
-		PreferencesManager.setValue(PreferencesManager.TXTUML_VISUALIZE_TXTUML_LAYOUT, layoutConfigs.values().stream()
-				.flatMap(c -> c.stream()).map(layout -> layout.getFullyQualifiedName()).collect(Collectors.toList()));
-
-		PreferencesManager.setValue(PreferencesManager.TXTUML_VISUALIZE_TXTUML_LAYOUT_PROJECTS,
-				layoutConfigs.values().stream().flatMap(c -> c.stream())
-						.map(layout -> layout.getJavaProject().getElementName()).collect(Collectors.toList()));
-
+	protected boolean exportDiagrams(Map<Pair<String, String>, List<IType>> layoutConfigs, List<IType> txtUMLLayout) {
 		for (Pair<String, String> model : layoutConfigs.keySet()) {
 			String txtUMLModelName = model.getFirst();
 			String txtUMLProjectName = model.getSecond();
@@ -116,23 +51,18 @@ public class PlantUMLVisualizeWizard extends Wizard {
 			String generatedFolderName = PreferencesManager
 					.getString(PreferencesManager.TXTUML_VISUALIZE_DESTINATION_FOLDER);
 
-			Map<String, String> layouts = new HashMap<String, String>();
-			layoutConfigs.get(model).forEach(
-					layout -> layouts.put(layout.getFullyQualifiedName(), layout.getJavaProject().getElementName()));
-
-			List<String> fullyQualifiedNames = txtUMLLayout.stream().map(IType::getFullyQualifiedName)
+			List<String> diagramNames = txtUMLLayout.stream().map(IType::getFullyQualifiedName)
 					.collect(Collectors.toList());
 			boolean saveSucceeded = SaveUtils.saveAffectedFiles(getShell(), txtUMLProjectName, txtUMLModelName,
-					fullyQualifiedNames);
+					diagramNames);
 			if (!saveSucceeded)
 				return false;
-
 			try {
-				this.checkNoLayoutDescriptionsSelected();
+				checkNoLayoutDescriptionsSelected();
 
 				IProgressService progressService = PlatformUI.getWorkbench().getProgressService();
 				PlantUmlExporter exp = new PlantUmlExporter(txtUMLProjectName, generatedFolderName,
-						fullyQualifiedNames);
+						diagramNames);
 
 				if (exp.hasSequenceDiagram()) {
 					progressService.runInUI(progressService, new IRunnableWithProgress() {
@@ -156,7 +86,7 @@ public class PlantUMLVisualizeWizard extends Wizard {
 						return true;
 					}
 				}
-			} catch (InvocationTargetException | InterruptedException e) {
+			} catch (InterruptedException | InvocationTargetException e) {
 				Logger.sys.error(e.getMessage());
 				return false;
 			}
@@ -164,16 +94,4 @@ public class PlantUMLVisualizeWizard extends Wizard {
 		return true;
 	}
 
-	private void checkNoLayoutDescriptionsSelected() throws InterruptedException {
-		if (selectPlantUmlPage.getPlantUmlLayouts().isEmpty()) {
-			boolean answer = Dialogs.WarningConfirm("No Layout descriptions",
-					"No diagrams will be generated using the current setup,"
-							+ " because no diagram descriptions are added." + System.lineSeparator()
-							+ "In order to have diagrams visualized, select a description from the wizard."
-							+ System.lineSeparator() + System.lineSeparator()
-							+ "Do you want to continue without diagram descriptions?");
-			if (!answer)
-				throw new InterruptedException();
-		}
-	}
 }
