+++ dev/plugins/hu.elte.txtuml.utils.eclipse/plugin.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+   <extension
+         point="org.eclipse.core.runtime.preferences">
+      <initializer
+            class="hu.elte.txtuml.utils.eclipse.preferences.PreferencesInitializer">
+      </initializer>
+   </extension>
+
+</plugin>
