+++ dev/plugins/hu.elte.txtuml.utils.eclipse/src/hu/elte/txtuml/utils/eclipse/Activator.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.export.papyrus;
+package hu.elte.txtuml.utils.eclipse;
 
 import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.ui.plugin.AbstractUIPlugin;
