+++ dev/plugins/hu.elte.txtuml.utils.eclipse/src/hu/elte/txtuml/utils/eclipse/preferences/PreferencesInitializer.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.export.papyrus.preferences;
+package hu.elte.txtuml.utils.eclipse.preferences;
 
 import org.eclipse.core.runtime.preferences.AbstractPreferenceInitializer;
